+++ src/client/admin/authentication/signin.component.ts
@@ -1,8 +1,8 @@
 import { Component } from '@angular/core';
 import { AuthenticationService } from './authentication.service';
-import { User } from './user.class';
+import { User } from '../user.class';
 import { UserStateService } from './user-state.service';
-import { ConfigService } from '../core/config.service';
+import { ConfigService } from '../../core/config.service';
 
 @Component({
 	templateUrl: './signin.component.html'
