+++ src/client/admin/authentication/user-state.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
 
-import { User } from './user.class';
+import { User } from '../user.class';
 
 /**
  *
