+++ src/client/admin/end-user-agreement/admin-create-eua.component.ts
@@ -5,7 +5,7 @@ import { Modal } from 'angular2-modal/plugins/bootstrap';
 
 import { EndUserAgreement } from './eua.class';
 import { EuaService } from './eua.service';
-import { AuthenticationService } from '../authentication.service';
+import { AuthenticationService } from '../authentication/authentication.service';
 import { ManageEuaComponent } from './manage-eua.component';
 import { AlertService } from '../../shared/alert.service';
