+++ src/client/admin/user-management/admin-list-users.component.ts
@@ -6,16 +6,16 @@ import * as _ from 'lodash';
 import { DialogRef, overlayConfigFactory } from 'angular2-modal';
 import { Modal } from 'angular2-modal/plugins/bootstrap';
 
-import { User } from './user.class';
-import { AdminService } from './admin.service';
-import { AuthenticationService } from './authentication.service';
+import { User } from '../user.class';
+import { AdminService } from '../admin.service';
+import { AuthenticationService } from '../authentication/authentication.service';
 import { Role } from './role.class';
 import { ExportUsersModalContext, ExportUsersModal } from './export-users.component';
-import { PagingOptions } from '../shared/pager.component';
-import { SortDisplayOption, SortDirection } from '../shared/result-utils.class';
-import { AlertService } from '../shared/alert.service';
-import { ExportConfigService } from '../shared/export-config.service';
-import { ConfigService } from '../core/config.service';
+import { PagingOptions } from '../../shared/pager.component';
+import { SortDisplayOption, SortDirection } from '../../shared/result-utils.class';
+import { AlertService } from '../../shared/alert.service';
+import { ExportConfigService } from '../../shared/export-config.service';
+import { ConfigService } from '../../core/config.service';
 
 @Component({
 	templateUrl: './admin-list-users.component.html'
