+++ src/client/admin/user-management/admin-user.module.ts
@@ -8,8 +8,8 @@ import { Ng2BootstrapModule } from 'ng2-bootstrap';
 import { AdminService } from '../admin.service';
 import { AdminCreateUserComponent } from './admin-create-user.component';
 import { AdminUpdateUserComponent } from './admin-edit-user.component';
-import { AdminListUsersComponent } from '../admin-list-users.component';
-import { ExportUsersModal } from '../export-users.component';
+import { AdminListUsersComponent } from './admin-list-users.component';
+import { ExportUsersModal } from './export-users.component';
 import { UpdateUserComponent } from './edit-user.component';
 import { UserSignupComponent } from './user-signup.component';
 import { UserService } from '../users.service';
