+++ src/client/app-routing.module.ts
@@ -1,17 +1,17 @@
 import { RouterModule } from '@angular/router';
 import { NgModule } from '@angular/core';
 import { AuthGuard } from './core/auth-guard.service';
-import { SigninComponent } from './admin/signin.component';
-import { UnauthorizedComponent } from './admin/unauthorized.component';
+import { SigninComponent } from './admin/authentication/signin.component';
+import { UnauthorizedComponent } from './admin/authentication/unauthorized.component';
 import { UserSignupComponent } from './admin/user-management/user-signup.component';
 import { UpdateUserComponent } from './admin/user-management/edit-user.component';
 import { InvalidResourceComponent } from './core/invalid-resource.component';
-import { InvalidCertificateComponent } from './admin/invalid-certificate.component';
+import { InvalidCertificateComponent } from './admin/authentication/invalid-certificate.component';
 import { AuditComponent } from './audit/components/audit.client.component';
 import { UserEuaComponent } from './admin/end-user-agreement/user-eua.component';
-import { NoAccessComponent } from './admin/no-access.component';
+import { NoAccessComponent } from './admin/authentication/no-access.component';
 import { LoggedInComponent } from './core/logged-in.component';
-import { InactiveComponent } from './admin/inactive.component';
+import { InactiveComponent } from './admin/authentication/inactive.component';
 
 @NgModule({
 	imports: [
