+++ src/client/app.module.ts
@@ -11,12 +11,13 @@ import { AdminModule } from './admin/admin.module';
 import { AccessCheckerModule } from './access-checker/access-checker.module';
 import { BootstrapModalModule } from 'angular2-modal/plugins/bootstrap';
 import { AuditModule } from './audit/audit.module';
-import { HeaderComponent, FooterComponent } from './core/core.component';
+import { HeaderComponent } from './core/header.component';
+import { FooterComponent } from './core/footer.component';
 import { LoggedInComponent } from './core/logged-in.component';
 import { InvalidResourceComponent } from './core/invalid-resource.component';
-import { AuthenticationService } from './admin/authentication.service';
+import { AuthenticationService } from './admin/authentication/authentication.service';
 import { AuthGuard } from './core/auth-guard.service';
-import { UserStateService } from './admin/user-state.service';
+import { UserStateService } from './admin/authentication/user-state.service';
 import { ConfigService } from './core/config.service';
 import { UtilModule } from './shared/util.module';
 import { ClientConfiguration } from './config/configurator';
