+++ src/client/core/help/core-help.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { OverviewHelpComponent } from './overview.component';
-import { AboutComponent } from '../../help/help.component';
+import { AboutComponent } from '../../help/about.component';
 import { HelpService } from '../../help/help.service';
 import { HelpTopic } from '../../help/help.class';
