+++ src/client/help/about.component.html
@@ -3,9 +3,9 @@
 	<div style="padding-top:20px;" class="row">
 
 		<div class="col-md-8 col-md-offset-2 jumbotron text-center">
-			<img class="logo" src="../../core/img/brand/logo.png" />
+			<img class="logo" src="../img/brand/logo.png" />
 			<br/>
-			<img class="logo-words" src="../../core/img/brand/text-light-bg.png" />
+			<img class="logo-words" src="../img/brand/text-light-bg.png" />
 			<br/>
 			<br/>
 			<strong>Version</strong>
