--- src/client/help/views/overview/overview.client.view.html
@@ -1,34 +1,0 @@
-<section>
-
-	<h1 class="page-header">Getting Started</h1>
-
-	<p class="lead">Before you start, it is important to understand the basic functionality and behavior of the application.</p>
-
-	<p>
-		Welcome to Application. Here you will find general documentation about the application as well as pointers to other resources that are available to you.
-		Navigate major topic areas in the help system using the links in the left column.
-		<!-- Will want to add a link to our help site and our email (which should both be configurable) -->
-	</p>
-
-	<div class="callout callout-info">
-		<h4>Browser Recommendation</h4>
-		<pre><i class="fa fa-3x fa-firefox"> Firefox</i>    <i class="fa fa-3x fa-chrome"> Chrome</i></pre>
-		<p>
-			Application is designed to work in all modern browsers, and it is optimized for Firefox and Chrome.
-		</p>
-	</div>
-
-	<div class="callout callout-info">
-		<h4>Looking for help?</h4>
-		<p>Contextual help is located throughout the system using these icons: <span data-asy-help=""></span></p>
-		<p>
-			You can also get back to this help section at any time by clicking on the help menu
-			<span class="fa-stack fa">
-				<i class="fa fa-circle-o fa-stack-2x"></i>
-				<i class="fa fa-question fa-stack-1x"></i>
-			</span>
-			in the top-right corner of the site.
-		</p>
-	</div>
-
-</section>
